begining of redirect

+++ src/shell_handling/execution_handling/execute_list.c
@@ -17,6 +17,7 @@
 # include "alias.h"
 # include "my.h"
 # include "sh.h"
+# include "redirect.h"
 
 static int start_exe(list_t *cmd, char ***env, shell_t *shell, int *status)
 {
@@ -31,18 +32,20 @@ static int start_exe(list_t *cmd, char ***env, shell_t *shell, int *status)
 }
 
 static int execute_command(list_t *cmd, shell_t *shell, char ***env,
-	int *status)
+			int *status)
 {
 	int	pid_son = 0;
-	list_t	*next_cmd = NULL;
+	list_t	*prev_cmd = NULL;
 
 	if (cmd == NULL)
 		return (0);
 	pid_son = start_exe(cmd->next[CMD], env, shell, status);
 	if (pid_son == -1)
 		return (-1);
-	next_cmd = cmd->next[SEPARATOR];
-	*status = execute_command(next_cmd, shell, env, status);
+	prev_cmd = cmd->prev;
+	*status = execute_command(prev_cmd, shell, env, status);
+	if (status != 0)
+		return (*status);
 	waitpid(pid_son, status, 0);
 	kill(pid_son, *status);
 	return (signal_handler(*status));
@@ -50,10 +53,14 @@ static int execute_command(list_t *cmd, shell_t *shell, char ***env,
 
 int	execute_list(shell_t *shell, char ***env, int *status)
 {
-	execute_command(shell->cmd, shell, env, status);
-	if (*status == -1) {
-		*status = 1;
-		return (1);
+	list_t	*tmp = shell->cmd;
+
+	while (tmp->next[SEPARATOR] != NULL) {
+		init_pipe(tmp);
+		tmp = tmp->next[SEPARATOR];
 	}
-	return (0);
+	*status = execute_command(tmp, shell, env, status);
+	if (*status == -1)
+		*status = 1;
+	return (*status);
 }
